pyproject.toml
[project]
name = "trellis"
version = "0.1.0"
description = "A lightweight data I/O adapter for data pipelines."
readme = "README.md"
authors = [
    { name = "Matthew McCabe", email = "mpmbq2@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pyarrow",
    "fsspec",
    "polars",
    "pandas",
    "duckdb",
    "ibis-framework",
    "pyarrow_hotfix",
]

[project.optional-dependencies]
all = []

[project.scripts]
trellis = "trellis:main"

[dependency-groups]
dev = [
    "pytest",
    "ruff",
    "mypy",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
